requirements.txt
pystray>=0.19.5
pynput>=1.7.6
Pillow>=10.0.0
fastmcp>=2.9.0
pyperclip>=1.8.2
